ドメインにガード条件追加

+++ examination2-app-server/src/main/kotlin/org/contourgara/examination2gara/infrastructure/BookEntity.kt
@@ -5,7 +5,7 @@ import org.contourgara.examination2gara.domain.Book
 /**
  * BookEntity は、本情報をデータベースとのマッピングをするためのクラスです。
  */
-data class BookEntity(val id: Int, val title: String, val author: String, val publisher: String, val price:Int) {
+data class BookEntity(val id: Long, val title: String, val author: String, val publisher: String, val price:Int) {
   /**
    * ドメインオブジェクトへと変換します。
    *
