+++ examination2-app-server/src/main/kotlin/org/contourgara/examination2gara/infrastructure/BookMapper.kt
@@ -9,5 +9,5 @@ interface BookMapper {
   fun findAll(): List<BookEntity>
 
   @Select("SELECT id, title, author, publisher, price FROM books WHERE id = #{id}")
-  fun findById(id: Int): BookEntity?
+  fun findById(id: Long): BookEntity?
 }
