+++ examination2-app-server/src/main/kotlin/org/contourgara/examination2gara/infrastructure/BookRepositoryImpl.kt
@@ -18,6 +18,6 @@ class BookRepositoryImpl(
   }
 
   override fun findById(id: String): Book? {
-    return bookMapper.findById(id.toInt())?.toModel()
+    return bookMapper.findById(id.toLong())?.toModel()
   }
 }
